+++ CHANGELOG.md
@@ -87,6 +87,7 @@
 - New #964: Add `QueryBuilderInterface::replacePlaceholders()` method (@Tigrov)
 - Enh #879: Rename `getLastInsertID()` method in `ConnectionInterface` to `getLastInsertId()` (@vjik)
 - New #967: Add `FOR` clause to query (@vjik)
+- Chg #972: Change in query "distinct" flag type from `bool|null` to `bool` (@vjik)
 
 ## 1.3.0 March 21, 2024